requirements.txt
astroid==2.4.2
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.2
click==7.1.2
decorator==4.4.2
dnspython==2.0.0
flake8==3.8.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
infinity==1.5
intervals==0.9.0
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.2.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
toml==0.10.1
validators==0.18.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.4
